Data_Generator/requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
python-dotenv==1.0.1
influxdb-client==1.48.0
flask_apscheduler==1.13.1